% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_dependencies.R
\name{session_dependencies}
\alias{session_dependencies}
\title{session_dependencies}
\usage{
session_dependencies(exclude_base = TRUE)
}
\arguments{
\item{exclude_base}{should base and recommended packages be excluded from dependencies?}
}
\description{
session_dependencies
}
