% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_snapshot.R
\name{library_snapshot}
\alias{library_snapshot}
\title{library_snapshot}
\usage{
library_snapshot(path = "rlib")
}
\arguments{
\item{path}{path to which to make copy to}
}
\value{
Returns the paths from which packages were copied
}
\description{
Make a local copy of the currently used packeges from the currently used library.
}
\examples{
\dontrun{

library_snapshot(path = tempdir())

}


}
