% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_snapshot.R
\name{library_snapshot}
\alias{library_snapshot}
\title{library_snapshot}
\usage{
library_snapshot(path = NULL)
}
\arguments{
\item{path}{path to which to copy packages to; if NULL}
}
\value{
Returns the paths from which packages were copied
}
\description{
Make a local copy of the currently used packages within the session
from the currently used library.
}
\examples{
\dontrun{


library_snapshot()

library_snapshot(path = tempdir())


}


}
