% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{package_dependencies}
\alias{package_dependencies}
\title{package_dependencies}
\usage{
package_dependencies(pkg, env = NULL, dependency_type = c("depends",
  "imports", "linkingto"))
}
\arguments{
\item{pkg}{the package to resolve dependencies for}

\item{env}{an environment to store results in, is primarily used for recursion}

\item{dependency_type}{the dependency types to include, e.g.
"depends", "imports", "linkingto", "suggests" -- will leave out suggests by
default}
}
\description{
package_dependencies
}
